requirements.txt
flask
flask-cors
python-dotenv
gunicorn
psycopg2-binary
